refactor(reports): extract project detection to reports_projects.go
